Fix to allow for better running when not focused tab
Changes allow for quest progress and ore earned to update more accurately when tab not focused.

Update Game.Logic to keep track of time since last time Game.Logic was run then update quest progress and amoutn of ore earned based onf time passed instead of always updating the same amount.

Updated Game.calculateRemainingQuest to accept an input which is the amount of time which as passed in ms.